Initial attempt at using ibeacon UUIDs for ble_rssi.

+++ esphome/components/esp32_ble_tracker/esp32_ble_tracker.h
@@ -52,8 +52,9 @@ class ESPBLEiBeacon {
   ESPBTUUID get_uuid() { return ESPBTUUID::from_raw(this->beacon_data_.proximity_uuid); }
 
  protected:
+  // See 'Artwork and Specifications' @ https://developer.apple.com/ibeacon/
   struct {
-    uint8_t sub_type;
+    uint16_t sub_type;
     uint8_t proximity_uuid[16];
     uint16_t major;
     uint16_t minor;
